+++ addOns/scripts/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- The Scripts output panel will now use the work panel font defined in Display settings (Issue 8065).
 
 ## [40] - 2023-09-11
 ### Changed